+++ src/orphan_txns.h
@@ -83,7 +83,7 @@ class COrphanTxns {
     /** Get extra transactions needed by block's reconstruction */
     CompactExtraTxnsVec getCompactExtraTxns() const;
     /** Limit a number of orphan transactions size */
-    unsigned int limitTxnsSize(unsigned int nMaxOrphanTxnsSize, bool fSkipRndEviction=false);
+    unsigned int limitTxnsSize(uint64_t nMaxOrphanTxnsSize, bool fSkipRndEviction=false);
     /** Collect dependent transactions which might be processed later */
     std::vector<TxInputDataSPtr> collectDependentTxnsForRetry();
     /** Collect txn's outpoints which will be used to find any dependant orphan txn */
